BOOK-19 bug-fix-test/users-security-test

+++ src/main/java/bookstore/javabrightbrains/controller/CartController.java
@@ -4,12 +4,14 @@ import bookstore.javabrightbrains.dto.cart.CartItemUpdateRequestDto;
 import bookstore.javabrightbrains.dto.cart.CartItemRequestDto;
 import bookstore.javabrightbrains.dto.cart.CartResponseDto;
 import bookstore.javabrightbrains.service.CartService;
+import bookstore.javabrightbrains.service.JwtSecurityService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -21,8 +23,10 @@ import static bookstore.javabrightbrains.utils.Constants.USER_BASE_URL;
 @RequiredArgsConstructor
 @Tag(name = "Cart Controller", description = "APIs for managing manages the user's shopping cart")
 public class CartController {
-
-    private final CartService cartService;
+    @Autowired
+    private  CartService cartService;
+    @Autowired
+    private JwtSecurityService jwtSecurityService;
 
     @GetMapping
     @Operation(summary = "Get the user's cart")
@@ -32,6 +36,7 @@ public class CartController {
             @ApiResponse(responseCode = "404", description = "Cart not found")
     })
     public ResponseEntity<CartResponseDto> getCart(@PathVariable Long userId) {
+        jwtSecurityService.validateUserAccess(userId);
         CartResponseDto cartResponseDto = cartService.getCart(userId);
         return ResponseEntity.ok(cartResponseDto);
     }
@@ -45,6 +50,7 @@ public class CartController {
             @ApiResponse(responseCode = "404", description = "Book not found")
     })
     public ResponseEntity<String> addToCart(@PathVariable Long userId, @Valid @RequestBody CartItemRequestDto cartItemRequestDto) {
+        jwtSecurityService.validateUserAccess(userId);
         cartService.addToCart(userId, cartItemRequestDto);
         return ResponseEntity.status(201).build();
     }
@@ -60,6 +66,7 @@ public class CartController {
     })
     public ResponseEntity<String> updateCartItem(@PathVariable Long userId, @PathVariable Long cartItemId,
                                                  @Valid @RequestBody CartItemUpdateRequestDto cartItemUpdateRequestDto) {
+        jwtSecurityService.validateUserAccess(userId);
         cartService.updateCartItem(userId, cartItemId, cartItemUpdateRequestDto);
         return ResponseEntity.status(204).build();
     }
@@ -73,6 +80,7 @@ public class CartController {
             @ApiResponse(responseCode = "404", description = "Item does not belong to the user")
     })
     public ResponseEntity<String> deleteCartItem(@PathVariable Long userId, @PathVariable Long cartItemId) {
+        jwtSecurityService.validateUserAccess(userId);
         cartService.deleteCartItem(userId, cartItemId);
         return ResponseEntity.status(204).build();
     }
